fix: Stop page rendering if user is not logged in
If a non-logged-in user attempted to go to the /add endpoint, the app would render part of the
page, then display a popup saying "Please log in to view this page", and the user would be
directed to the login screen.  This was kind of ugly, so I changed it so that nothing is rendered
on any pages, unless the user is logged in.