fix: Fix router problem
When a user navigated to /questions/question_id it was being stacked on top of the /home
URL, as I had forgotten a  backslash. This commit fixes that.  Now home is at localhost/home
and questions are at localhost/questions/question_id.  Also updated the readme, just a few words
about how to run tests (as there seem to be 2 commands that work).